requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.16
langchain-openai==0.0.5
chromadb==0.4.22
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
PyPDF2==3.0.1
tiktoken>=0.5.2,<0.6.0
openai>=1.10.0,<2.0.0
docx2txt==0.8
sentence-transformers==2.5.1 